harmony-matcher/backend_fastapi/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
httpx>=0.27.0
twilio>=9.0.0
anthropic>=0.25.0
openpyxl>=3.1.2


